Organize pages and finish footer.
