+++ Frontend/src/Pages/AboutPage/AboutPage.tsx
@@ -1,0 +1,9 @@
+function AboutPage() {
+  return (
+    <>
+      <h1 className="title-style">About the Developers</h1>
+    </>
+  );
+}
+
+export default AboutPage;
