+++ Frontend/src/main.tsx
@@ -1,13 +1,13 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import LandingPage from "./LandingPage/LandingPage";
+import LandingPage from "./Pages/LandingPage/LandingPage";
 import { BrowserRouter, Route, Routes } from "react-router";
-import Header from "./Header/Header";
-import PomodoroPage from "./PomodoroPage/PomodoPage";
-import ErrorPage from "./ErrorPage/ErrorPage";
-import Footer from "./Footer/Footer";
-
+import Header from "./common/Header/Header";
+import PomodoroPage from "./Pages/PomodoroPage/PomodoPage";
+import ErrorPage from "./Pages/ErrorPage/ErrorPage";
+import Footer from "./common/Footer/Footer";
+import AboutPage from "./Pages/AboutPage/AboutPage";
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <BrowserRouter>
@@ -16,6 +16,7 @@ createRoot(document.getElementById("root")!).render(
         <Route path="/" element={<LandingPage />} />
         <Route path="/Pomodoro" element={<PomodoroPage />} />
         <Route path="*" element={<ErrorPage />} />
+        <Route path="/AboutUs" element={<AboutPage />} />
       </Routes>
       <Footer />
     </BrowserRouter>
